requirements.txt
# packages which are needed for the code to run

streamlit
openai
faiss-cpu
python-dotenv
langdetect
PyMuPDF
python-docx

### you do this to install all of them : pip install -r requirements.txt
